TECH-7: Fixing button logic
